# jaxrs-spec/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Che Feature API",
    "version" : "1.0.0",
    "title" : "Che Feature API",
    "contact" : {
      "email" : "skabashniuk@redhat.com"
    },
    "license" : {
      "name" : "Eclipse Public License v1.0",
      "url" : "http://www.eclipse.org/legal/epl-v10.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/skabashniuk/Che/1.0.0",
  "schemes" : [ "https" ],
  "paths" : {
    "/service" : {
      "get" : {
        "summary" : "Searches CheService by list of ids",
        "description" : "Searches  CheService by list of ids, example ?id=io.typefox.theia-ide.che-service:1.2.3&org.eclipse.che.svn:4.2.3",
        "operationId" : "getServiceByIdList",
        "produces" : [ "application/x-yaml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "list if CheService ids.",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "The CheService successfully fetched",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CheService"
              }
            }
          },
          "500" : {
            "description" : "500 Internal server error occurred",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "summary" : "Create a new CheService",
        "description" : "Create a new CheService",
        "operationId" : "addCheService",
        "consumes" : [ "application/json", "application/x-yaml" ],
        "produces" : [ "application/json", "application/x-yaml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "CheServiceItem",
          "description" : "CheFeature item to add",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CheService"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "The CheService successfully created",
            "schema" : {
              "$ref" : "#/definitions/CheService"
            }
          },
          "400" : {
            "description" : "400 Server receives invalid input parameter",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "403 The user does not have access to update resource",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "409" : {
            "description" : "409 Operation could not be performed because of conflict with prior state.",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/service/{name}/{version}" : {
      "get" : {
        "summary" : "Get CheService by name and version",
        "description" : "Get CheService by name and version",
        "operationId" : "getServiceByNameAndVersion",
        "produces" : [ "application/x-yaml", "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "CheService name.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "version",
          "in" : "path",
          "description" : "CheService version.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "The CheService successfully fetched",
            "schema" : {
              "$ref" : "#/definitions/CheService"
            }
          },
          "500" : {
            "description" : "500 Internal server error occurred",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Removes CheService by name and version",
        "description" : "Removes CheService by name and version",
        "operationId" : "deleteServiceByNameAndVersion",
        "produces" : [ "application/x-yaml", "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "Service name.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "version",
          "in" : "path",
          "description" : "Service version.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "The CheService successfully removed"
          },
          "403" : {
            "description" : "403 The user does not have access to update resource",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "404 The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "500 Internal server error occurred",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/feature" : {
      "get" : {
        "summary" : "Searches CheFeature by list of ids",
        "description" : "Searches CheFeature by list of ids. For example, ?id=io.typefox.theia-ide.che-service:1.2.3&org.eclipse.che.svn:4.2",
        "operationId" : "searchFeatureByListOfId",
        "produces" : [ "application/x-yaml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "list if CheService ids.",
          "required" : true,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "The CheFeature successfully fetched",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CheFeature"
              }
            }
          },
          "500" : {
            "description" : "500 Internal server error occurred",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "summary" : "Create a new CheFeature",
        "description" : "Create a new CheFeature",
        "operationId" : "addFeature",
        "consumes" : [ "application/json", "application/x-yaml" ],
        "produces" : [ "application/json", "application/x-yaml" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "CheFeatureItem",
          "description" : "CheFeature item to add",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CheFeature"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "The CheFeature successfully created",
            "schema" : {
              "$ref" : "#/definitions/CheFeature"
            }
          },
          "403" : {
            "description" : "403 The user does not have access to update resource",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "404 The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "409" : {
            "description" : "409 Operation could not be performed because of conflict with prior state.",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "500 Internal server error occurred",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/feature/{name}/{version}" : {
      "get" : {
        "summary" : "Get CheFeature by name and version",
        "description" : "Get CheFeature by name and version",
        "operationId" : "searchFeatureByNameAndVersion",
        "produces" : [ "application/x-yaml", "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "CheFeature name.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "version",
          "in" : "path",
          "description" : "CheFeature version.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "The CheFeature successfully fetched",
            "schema" : {
              "$ref" : "#/definitions/CheFeature"
            }
          },
          "500" : {
            "description" : "500 Internal server error occurred",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "delete" : {
        "summary" : "Removes CheFeature by name and version",
        "description" : "Removes CheFeature by name and version",
        "operationId" : "deleteFeatureByNameAndVersion",
        "produces" : [ "application/x-yaml", "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "description" : "CheFeature name.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "version",
          "in" : "path",
          "description" : "CheFeature version.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "The CheFeature successfully removed"
          },
          "403" : {
            "description" : "403 The user does not have access to update resource",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "404 The specified resource was not found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "409" : {
            "description" : "409 Operation could not be performed because of conflict with prior state.",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "500 Internal server error occurred",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "CheService" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TypeMeta"
      }, {
        "type" : "object",
        "required" : [ "metadata", "spec" ],
        "properties" : {
          "metadata" : {
            "$ref" : "#/definitions/ObjectMeta"
          },
          "spec" : {
            "$ref" : "#/definitions/CheServiceSpec"
          }
        }
      } ]
    },
    "CheServiceSpec" : {
      "type" : "object",
      "required" : [ "containers", "version" ],
      "properties" : {
        "version" : {
          "type" : "string",
          "example" : "1.0.0"
        },
        "containers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Container"
          }
        }
      }
    },
    "Container" : {
      "type" : "object",
      "required" : [ "image" ],
      "properties" : {
        "image" : {
          "type" : "string",
          "example" : "eclipse/che-theia:nightly"
        },
        "env" : {
          "type" : "array",
          "description" : "List of environment variables to set in the container. Cannot be updated.",
          "items" : {
            "$ref" : "#/definitions/EnvVar"
          }
        }
      }
    },
    "ResourceRequirements" : {
      "type" : "object",
      "required" : [ "requests" ],
      "properties" : {
        "requests" : {
          "type" : "array",
          "description" : "Requests describes the minimum amount of compute resources required.",
          "items" : {
            "$ref" : "#/definitions/ResourceList"
          }
        }
      }
    },
    "ResourceList" : {
      "type" : "object"
    },
    "EnvVar" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "description" : "Name of the environment variable. Must be a C_IDENTIFIER."
        },
        "value" : {
          "type" : "string",
          "description" : "Variable references $(VAR_NAME) are expanded using the previous defined environment variables in the container and any service environment variables. If a variable cannot be resolved, the reference in the input string will be unchanged. The $(VAR_NAME) syntax can be escaped with a double $$, ie: $$(VAR_NAME). Escaped references will never be expanded, regardless of whether the variable exists or not. Defaults to \"\"."
        }
      },
      "description" : "EnvVar represents an environment variable present in a Container."
    },
    "CheFeature" : {
      "allOf" : [ {
        "$ref" : "#/definitions/TypeMeta"
      }, {
        "type" : "object",
        "required" : [ "metadata", "spec" ],
        "properties" : {
          "metadata" : {
            "$ref" : "#/definitions/ObjectMeta"
          },
          "spec" : {
            "$ref" : "#/definitions/CheFeatureSpec"
          }
        }
      } ]
    },
    "CheFeatureSpec" : {
      "type" : "object",
      "required" : [ "services", "version" ],
      "properties" : {
        "version" : {
          "type" : "string",
          "example" : "1.0.0"
        },
        "services" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CheServiceReference"
          }
        }
      }
    },
    "CheServiceReference" : {
      "type" : "object",
      "required" : [ "name", "version" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "o.typefox.theia-ide.che-service.mainpod"
        },
        "version" : {
          "type" : "string",
          "example" : "51.233"
        },
        "services" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CheServiceParameter"
          }
        }
      }
    },
    "CheServiceParameter" : {
      "type" : "object",
      "required" : [ "name", "value" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "THEIA_PLUGINS"
        },
        "value" : {
          "type" : "string",
          "example" : "org.eclipse.che.terminal@latest"
        }
      }
    },
    "TypeMeta" : {
      "type" : "object",
      "required" : [ "apiVersion", "kind" ],
      "properties" : {
        "kind" : {
          "type" : "string",
          "example" : "CheService",
          "description" : "Kind is a string value representing the REST resource this object represents."
        },
        "apiVersion" : {
          "type" : "string",
          "example" : "che.eclipse.org/v1",
          "description" : "APIVersion defines the versioned schema of this representation of an object"
        }
      }
    },
    "ObjectMeta" : {
      "type" : "object",
      "required" : [ "name" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "org.eclipse.che.theia.git",
          "description" : "Object name. Name must be unique."
        },
        "labels" : {
          "type" : "object",
          "description" : "Map of string keys and values that can be used to organize and categorize (scope and select) objects. May match selectors of replication controllers and services. More info: http://kubernetes.io/docs/user-guide/labels",
          "additionalProperties" : {
            "type" : "string"
          }
        }
      }
    },
    "Error" : {
      "type" : "object",
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string"
        }
      }
    }
  },
  "responses" : {
    "BadRequestException" : {
      "description" : "400 Server receives invalid input parameter",
      "schema" : {
        "$ref" : "#/definitions/Error"
      }
    },
    "ForbiddenException" : {
      "description" : "403 The user does not have access to update resource",
      "schema" : {
        "$ref" : "#/definitions/Error"
      }
    },
    "NotFoundException" : {
      "description" : "404 The specified resource was not found",
      "schema" : {
        "$ref" : "#/definitions/Error"
      }
    },
    "ConflictException" : {
      "description" : "409 Operation could not be performed because of conflict with prior state.",
      "schema" : {
        "$ref" : "#/definitions/Error"
      }
    },
    "ServerException" : {
      "description" : "500 Internal server error occurred",
      "schema" : {
        "$ref" : "#/definitions/Error"
      }
    }
  }
}